Recommend exclude sensor.xiaomi_map from recorder
I was wondering why my database was getting so big. Especially why it was growing so ridiculously fast sometimes - and sometimes not.
Now I realised that it is the map data which makes the database so big. Of course ... new map data every 5 seconds for about an hour (~ 800 MB) every two days 😆